实现了商品表联合分类表查询，但是不完善，还需要创建单位表联合单位表查询

+++ src/main/java/com/example/mybatisplus/model/domain/Multibarcode.java
@@ -31,7 +31,7 @@ public class Multibarcode extends Model<Multibarcode> {
     @TableId(value = "id", type = IdType.AUTO)
     private Integer id;
 
-    private Integer comId;
+    private Long comId;
 
     private String barcode;
